Move delete error handling into errors.go
